New Clock App in SVG, updated schedule-function to return object
New Clock app will test ideas in SVG animation and interactivity.
Schedule-Function.js was changed to return an object instead of
just the value of continueAnimation. This allows user code to
update different data, In this case the timeout is passed to allow
when the next animation to be scheduled correctly.