SVG Animation broke for some reason, code altered to fix it.
This is a weird issue. There are 5 animateTransforms on this page
linked together, kind of referencing one another. If the faster/slower
function is run against all in a loop (normal way) the animation
freezes. If I break in the loop and do each separately, still in the
same function, it works and no problem.
To fix if first added a new animateTransform, connected as well and
then, don't cycle that one through the loop, as it isn't used for
anything else. The animation works as expected again, but there is a
pause about as long as the shortest "dur", so the sixth animation was
set at 0.1s and the delay to start animating is not noticable anymore.

Very weird, but fixed for now.