Covid-19 Stats Graph, starting code
The current end goal is to show time series stats for all states:
 1. infections/100,000 persons
 2. deaths/100,000 persons
 3. percent rise in cases over various periods of time.

The first two numbers always go up, so it is important to show
where the current monemtum of cases is happening, or happened in
the past. Some states or rural areas had low infection rates early
on, but over time surprising catch-ups took place.

The third number shows again what is necessary to get out of danger:
slower growth of cases. But there are two forces here. One is that
some states are testing more, but the percent of new cases over testing
has gone down. Some states have a goal of getting the percent positive
to low single figures. The second is hot-spots where percent increase
is over 10% per week. These states are marching up the list of the
first two stats, so they are easy to pick out of the pack.