Added new color theory: dual-compliment
As an example of how you add a new color-theory, I added
dual-compliment. A complimentary color is a color pair
where the hue angle is 180 degrees apart. A split compliment adds
and subtracts a given angle from the complimentary color producing
a total of three colors (one given). If the chosen angle is
60 degrees, the resulting three colors are a triadic group each
separated from the others by 120 deg.

The new theory, I'm calling dual-compliment, producing four colors,
or two pairs of complimentary colors. The given angle sets how
far apart the original color is from the secondary complimentary
group. If this angle is 90 deg, then the four colors form a tetradic group.

Adding a new theory requires the following steps:

1) add a new HTML select option to choose the theory:
   in this case the value "dual-compliment" was used.
2) Use this value to add a new TheoryGroup:

Theory["dual-compliment"] = new TheoryGroup("dual-compliment",[3,1,2,4]);

The [3,1,2,4] maps the theory group to the swatch set for the group.

3) finally, just add a little code to derive the group given an input
   of the original color and the angle, if used (a new case):

   case "dual-compliment":
        if (h >= 180) {
            hueCompliment = h - 180;
        }
        else
        if (h < 180) {
            hueCompliment = 180 + h;
        }
2 -->   TheoryColors[TheoryColors.length] = toHsl(hueCompliment,s,l);
        dualHue = (360 + h + angle)%360;
        dualCompliment = (180 + dualHue)%360;
3 -->   TheoryColors[TheoryColors.length] = toHsl(dualHue,s,l);
4 -->   TheoryColors[TheoryColors.length] = toHsl(dualCompliment,s,l);
        break;

Note that the original color is already added to the group, so adding
three new colors derived from the original makes up the four colors.