Explored optimization of Ackermann's Function
Discovered, I now have an algo which runs in ~O(1) time for
any set of function values which are small enougth to compute
the final value.

Also found out that tcl can out of the box compute integers to
2^1023, so can compute ackermanFn(3,1020) = 2^1023 -3.

Also, Tcl is stackless, but still has a recursion limit of 1000,
but this can be set dynamically per interp.