Revised dashoffset calculation on tartan threads
The new algo conforms with The Scottish Register of Tartans images.

Im not sure if there is an exact way to do it, but it should be
easier to notice actual mistakes on input if the images matches a
known control.
I also noticed that adjusting options for dashoffsets might be an interesting
feature that could add to pattern creation not meant to be a tartan.