Added vastly simplified algo which does not work.
It is interesting though in how it fails. for the 25-5-3 problem
it takes 10 heats to get the top three horses. The followers
are not tracked but put into a layer representing their place.
But this neglects to update when a later race by the winner pushes
them further down in layers.