Algo for finding non-paired array element problem
The algorithm "appears" to be O(N), but apparently using splice which
removes or adds elements at any index position is far from O(N).

Page: 001-day-final-3.html
Again, the algo works but trouble is apparent after 100k elements at
time complexity goes through the roof. At 10M elements, the running
time was over an hour!

Page: 001-day-final-4.html
The algorithm is essentailly the same, however the original array is
never modified. Instead the buckets contain the index of the element,
the bucket number is the remainder(%). At each pass through the
remaining elements only one bucket is passed on (the one with an odd
number of elements). This just requires a little indirection to convert
this index to the original array index.

However there is a bug in final-4 given this array:

[
    9655001,
    6802151,
    1096451,
    9655001,
    9655001,
    4519871,
    9655001,
    6802151,
    6802151
]

The number of buckets chosen goes 5, 3, 5, 3... Once the number of
iterations gets past the set limit (N*1.25), the final step only
works if every element is the same (4519871 in this case). So the
final-5 file will fix this detail, the buckets values should go like
this: 5,3,2,2,2.... At every 2, the remaining items are checked
bit by bit from 10, 100, 1000, etc.