Finished with composition table 8x8 and wow, the comp table
exactly matches "Chaos and Fractals" pg 231, table 5.20.

Note, the starting point is the normal cartesian grid rotated 180 deg
around the x-axis. Y increases as you move down. This means the image
also starts upside down and transforms which look like rotations, rotate
clockwise instead of counter-clockwise. The result of any transform
is the rotated version of the normal cartesian system.

The next task is adding in scaling 3 transforms by 1/2 and translations of
(0,0),(1/2,0),(0,1/2) and (1/2,1/2), filling 3 of 4 slots. There should
be 512 different options, each a template for a Sierpinski Gasket family
member.