Moved files to new types to avoid execution by naviserver
Some tcl files under pageroot are regular tcl files which are
not indended to be executed by naviserver. These are renamed
with the extension .nstcl. Note that the tclsh and naviserver
have not problem executing the files with this new ending, so
it wont interfere with functionality.

Also worked on find-permuted and compared how it works with
a rather lengthy regular expression which sort-of with a lot of
extra hints "works". This shows an important limitation for
regexps.