Confirmed and fixed problem with chaining animations
SVG uses SMIL animations which have a limited facility
to fix the geometry of the things you are animating.

When an animation finishes you can "freeze", which
means the state is that found in the last frame of the
animation, or you can return to the initial state, or
discard the objects.

A simple example problem is to swap the location of two
objects, if you repeat the process, taking into account
that the two objects are now in a different location and
you will need to apply the inverse transform, what you will
find is that the object first reverts to the start of
the previous animation and now the inverse transform does
something not helpful.