Improved graphics of possible ellipse paths etc
Etcetera = The SVG Path description for an arc includes
two points and a degree input in addition to sweep and direction
flags.

The degree is easy to calculate, just the slope of the line, but this
is only one possible, and usually not the best one to use.

Note that it the line joining the two endpoints has a slope and
if if dx and dy are both non-zero, you can draw a triangle with the
connecting line being the hypotenuse. If the degree/angle used in
the path is 0 or 90 (or equally 180, 270), the elliptic path passes
through all three endpoints of the "inscribed" triangle. One of these
is pretty much guaranteed to be way too big. If the short side is dy,
the angle of 90 will create a very tall ellipse.