Just tweaked a little of the side graph display.
The initial reason for the side graph was to choose,
the elliplse with the smallest area, and the lower part
was a rotated version of the top. Now I think I should
show rotations of the ellipse main axis which has a much
bigger and easier to predict effect on the size of the
ellipse and thus the length of the curve.