Code straightening after successful animation app:
1. Added timeout/interval to page state.  Huge bonus to
adding the kink of a pagestate object which provides a
simple way to sync urlvars, formvars and key code vars...
and default values.
2. Fixed weird <nobr> issue. At some point I accidentally removed
the closing tag and chrome was working overtime adding nobr tags
everywhere.
3. Hooked the 'strategy' button (in this case shuffle strategy) into
the anim code. In the future this button action will update as chained
steps in the animation.
4. Something else new with this animation code is the improvement to
add animation actions (in this case "next","prev", but could be more),
but this is a big help since it reuses the animation framework and
keeps the interface uncluttered with invalid options.

This commit is only partial cleanup, prior to changeing a few
var names (probably numPrisoners becomes numItems???).