Example of using separate tcl interp for resource management
Tcl allows for creating additional interps to separate untrusted of
potentially dangerous code from the main application. The new interp
only has resources given to it and can't introspect the parent interp.

In the case of calculating with highly recursive functions, stack
depth could be set to a lower or higher level of stack depth, or in the
case here to just the right level.

The code is also an example of using a tcl list to accumulate the fib
numbers and provide for both a simple use of memoization/dynamic programming, as well as the storage of the full sequence for display as a visualization.